hw4/requirements.txt
###### Requirements with Version Specifiers ######
Cython >= 0.25.2
graphviz >= 0.6
matplotlib >= 2.0.0
nltk >=3.2.2
numpy >=1.12.1
Pillow >=4.1.0
scikit-image >= 0.13.0
scikit-learn >= 0.18.1
scipy >= 0.19.0
word2vec >= 0.9.1
adjustText >= 0.6.1
